fix: use BIGINT for bank_account_number to avoid overflow

+++ src/main/java/com/example/agricole/entity/BankAccount.java
@@ -11,6 +11,6 @@ public class BankAccount extends FinancialAccount {
     private Bank bankName;
     private int bankCode;
     private int bankBranchCode;
-    private int bankAccountNumber;
+    private long bankAccountNumber;
     private int bankAccountKey;
 }
